+++ src/kiosk_portal/services/role_service.ts
@@ -1,0 +1,9 @@
+import { HOST } from "../../@app/constants/host";
+import request from "../../@app/utils/http_client";
+
+//Role
+export const getListRoleService = async () => {
+    const response = await request.get(`${HOST}/v1/roles`);
+    return response.data;
+  };
+
